✨ (#77, #92) API 연동 데이터  및 컴포넌트 분리
